# shared/openapi.yaml
openapi: 3.0.0
info:
  title: Notes API
  version: 1.0.0
  description: A simple notes API
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
  /notes:
    get:
      summary: Get all notes
      description: Get all notes
      tags:
        - notes
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesResponse'
    post:
      summary: Create a new note
      description: Create a new note
      tags:
        - notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notes/{id}:
    get:
      summary: Get a note by ID
      description: Get a note by ID
      tags:
        - notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a note
      description: Update a note
      tags:
        - notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a note
      description: Delete a note
      tags:
        - notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Note deleted successfully
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Note:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: string
          description: Unique identifier for the note
        title:
          type: string
          description: Title of the note
        content:
          type: string
          description: Content of the note
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    CreateNoteRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the note
        content:
          type: string
          description: Content of the note
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated title of the note
        content:
          type: string
          description: Updated content of the note
    NotesResponse:
      type: object
      required:
        - notes
      properties:
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message